Add support for Chebyshev polynomial of the first kind as basis.
Minterpy now supports Chebyshev polynomial (of the first kind)
basis by wrapping the one-dimensional Chebyshev polynomial of the first
kind from SciPy and allows for a multi-dimensional construction
via tensor product.
The new polynomial basis has a consistent interface
with the other Minterpy polynomial bases.

Currently, the following functionalities are supported for
instances of the `ChebyshevPolynomial` class:

- Evaluation on query points
- Transformation from and to the Lagrange basis
- Transformation from and to the Newton basis
- Transformation from and to the canonical basis
- Ordinary regression with the Chebyshev polynomial basis

The following are not yet supported and will be part of
separate commits:

- Definite integration
- Differentiation

Furthermore, the present update does not consider in-depth
any performance concerns in the evaluation or transformation
If such issues arise, separate issues may be opened.

This commit should resolve Issue #79.

Additional changes:

- The test suite for polynomials is extended to test the behaviors
  of Chebyshev polynomials.
- The test suite for transformations is refactored and extended
  to test the bheaviors of all supported transformations, including
  the ones involving Chebyshev polynomials.
- The test suite of the ordinary polynomial regression is extended
  to include testing the Chebyshev polynomial basis.
- The API documentation is updated to include the Chebyshev polynomial
  basis and the related transformations.