Introduce keyword-only arguments in several methods of multivariate polynomial ABC.
The signatures of several high-level public methods belonging
to multivariate polynomial ABC are modified to allow for keyword-only
arguments so the behavior of the undelrying functions can be modified
while respecting the expected common behavior of the methods accross
different concrete implementations.

The methods whose signature are modified are:

- `__call__()` of `MultivariatePolynomialABC`
- `_eval()` of `MultivariatePolynomialABC`
- `partial_diff()` (resp. `_partial_diff()`) of
  `MultivariatePolynomialSingleABC`
- `diff()` (resp. `_diff()`) of `MultivariatePolynomialSingleABC`
- `integrate_over()` (resp. (`_integrate_over()`) of
  `MultivariatePolynomialSingleABC`

This commit should resolve Issue #136.

Additional changes:

- For consistency, the private method `_eval()` of
  `MultivariatePolynomialABC` is now set to be a static method. The call
  to the method inside `__call__()` method is modified accordingly.
- The docstrings for the methods mentioned above are updated.