Fix issue with computing derivative term beyond specified order
When numba is deactivated, the test fails due to the attempt
to compute the derivative of monomial beyond the specified order.
When numba is activated, there is no strict array bound checking
so no error was raised.

Furthermore, a test from JIT-compiled combinations are removed
because the exception is only raised when numba is activated.

This commit should resolve Issue #146.