Refactor integration of monomial in polynomial classes.
The routines to integrate polynomial in each supported basis
have been refactored, moving the computations of the quadrature weights
to a separate module.
This is done to distinguish a bit more clearly the different layers of
abstraction in Minterpy. The polynomial classes should have
minimal interactions with the low-level implementation
details (that mostly operate on arrays).
The computation of quadrature weights exposes the upper layer to
many lower layer constructs and clutter the polynomial modules.

All polynomials definite integrations are now in the form of
coefficients multiplied by the quadrature weights (the detail of the latter
has been moved).

Additionally the `dds()` function in `minterpy.dds` is temporarily replicated
with a different interface to deal directly with low-level data structure
instead of an instance of `MultiIndexTree`.

There is no change in the code functionality.

This commit is part of resolution of Issue #114.