Multi-index set is now a read-only property of Grid
Instead of an instance attribute, an instance of `MultiIndexSet`
is now stored as a property (decorated by `@property`)
of a `Grid` instance.
The property is without setter (i.e., read-only).

Furthermore, a verification procedure is devised to process
the instance of `MultiIndexSet` passed to the constructor.
Additionally, tests are updated to check the behavior
of passing invalid type of multi-index set to the constructor.

This commit is part of resolution of Issue #71.