Refactor type checking in the verification module.
The function `check_type()` in the verification module
is refactored to use `isinstance()` instead of `issubclass()`.
The `expected_type` argument is now mandatory
when calling the function for clarity.

Furthermore, the function `check_type_n_values()` in favor
of using two functions: `check_type()` and `check_values()`
because `check_values()` only applies if the type is numerical.