Generating points is now a read-only property of Grid.
Instead of an instance attribute, an array of generating points
is now stored as a property (decorated by `@property`)
of a `Grid` instance.
The property is without a setter (i.e., read-only)

Furthermore, a verification procedure is devised to process
the array of generating points directly provided
to the `Grid` constructor.
The test suite has been subsequently updated.

This commit is part of the resolution of Issue #71.