Polynomial degree is now a read-only property of Grid.
The polynomial degree of a grid indicates the maximum polynomial
degree of one-dimensional polynomials in any dimension
that the grid can support.
A procedure to check if the available generating points
can support polynomials specified by the multi-index set
is made available via the helper function.

Several tests have been added to verify the behavior
of instances of the updated Grid.

This commit is part of the resolution of Issue #71.