Refactor the default constructor of Grid class.
The default constructor of the `Grid` class is refactored.
It now accepts as optional arguments: `generating_function`
and `generating_points`.
The generating function is the function to construct
generating points of the Grid. If not provided, then
the default of Leja-ordered Chebyshev-Lobatto generating
function is selected (as it was previously selected
albeit indirectly).
Generating points may be provided, but if no generating
function is provided then the grid may not be updated to
a higher degree.

The refactoring preserve the most basic call to construct
a `Grid` instance as it was before.

Furthermore:

- `generating_function` is now stored as a read-only
  property of a `Grid` instances and also a factor
  in determining instances equality in value.
- A method `expand_dim()` is introduced to instances
  of `Grid` to encapsulate the procedure of expanding
  the dimension of a `Grid` by `MultivariatePolynomialSingleABC`.
- Various internal methods and functions are introduced
  for the `Grid` class to verify and assign the arguments
  passed to the constructor.
- The test suite of the `Grid` class has been extended to
  test the behavior of the updated code.

This commit is part of the resolution of Issue #71.