Add generating function check to expand_dim() in Grid class.
The `expand_dim()` method now works only with instances whose
generating function to expand the dimension.
If there is no generating function available and the dimension
expansion is attemepted, an exception is raised.
However, if the target dimension stays the same, then
it does not matter if a generating function is available.

Additionally:

- a new property `has_generating_function()` is now available
  to quickly check if an instance has a generating function
  assigned to it.
- The test suite has been updated to reflect the new changes.

This commit should resolve Issue #151.