Add functionality for evaluating a function on a Grid.
An instance of `Grid` may now be used as a `Callable`.
The `__call__()` method of the `Grid` class evaluates
a given function on the unisolvent nodes of the grid
and returns the corresponding function values.

Additionally:

- Additional arguments to the function may be passed in
  the call to the `Grid` instance either via positional
  or keyword arguments.
- The test suite has been updated to reflect these changes.
- A guide on how to use this feature is added
  to the documentation.
- The not-yet-implemented method `apply_func()` has been removed
  in favor of `__call__()`.
- The docstring of the property `unisolvent_nodes` has been updated
  as well as the test suite.

This commit should resolve Issue #153 and part of the refactoring
of the `Grid` class (Issue #135).