Remove support for evaluating lower-dim. functions by Grid instance.
For simplicity, calling the Grid with a function or a callable
requires the function to be called
with a two-dimensional array with the number of columns
corresponds to the number of spatial dimension of the Grid instance.
The given function may or may not be called with an array
with more columns, but it is not the responsibility
of the `Grid` instance to deal with this.

The corresponding tests for this functionality
in the test suite have been deleted.