Update expand_dim() method of the Grid class.
The instance method `expand_dim()` of the `Grid` class
has been updated to allow a grid instance to be passed
as the `target_dimension`.
Passing a `Grid` instance means that the current instance
should be expanded to the dimension of the `Grid` instance as long
as the underlying generating function or points are compatible.

The relevant tests have been updated to ensure the correct
behavior introduced by these changes.