Generating points of a Grid may be of higher dimension.
Generating points with excess dimensions (i.e., more columns)
than what is required by the multi-index set may now be used
to construct a Grid instance.

This also means that expanding the dimension of a `Grid` instance
without a generating function is allowed as long as
the underlying generating functions allow it.

The test suite has been updated to include these changes.