Add merge() method to Grid class.
A new `merge()` method is added to the `Grid` class
to combine two `Grid` instances with a new multi-index set.

This merge method takes care of validating the compatibility
of generating functions and points between Grid instances,
and throws an error if they are not compatible.

This helps to simplify and centralize the logic needed
for merging grids.