Implement multiplication operation for Grid instances.
The dunder method `__mul__()` is now implemented
in the `Grid` class that allows `Grid` instances
to be multiplied with each other.

A multiplication between two `Grid` instances
returns a new `Grid` instance with a multi-index
set that is the product of the multi-index sets
of the operands.

The test suite has been updated to include
the changes. Finally, a how-to guide on
`Grid` multiplication has been added to the docs.

This commit should resolve Issue #154.