Implement union operation for Grid instances
Instances of `Grid` may now be unionized via
the `|` operator. The result is a new `Grid`
instance whose multi-index set is the union of
the index sets of the operands.

The test suite has been updated accordingly to reflect
the latest changes. Furthermore, a new How-To Guide
demonstrating the new feature is added to the docs.

This commit should resolve Issue #155.