Refactor the method to make a Grid instance complete.
The method `make_complete()` has been refactored to always return
a new instance of `Grid`. If the current instance is not complete,
then a complete multi-index set is first created and a new instance
of `Grid` is then created based on the new multi-index set.

Additionally:

- A new property `is_complete()` is available for `Grid` instances;
  it carries over from the underlying `MultiIndexSet` a property
  of the same name.
- The test suite has been updated to reflect the current changes.
- A How-To Guide demonstrating the usage  of `make_complete()`
  is added to the docs.

This commit should resolve Issue #157.