Refactor add_points() method of the Grid class.
The method `add_points()` of the `Grid` class has been
renamed to `add_exponents()` to be consistent with a method
of the same name of the `MultiIndexSet` class.
Calling the method on an instance with a set of exponents
adds the exponents to the underlying multi-index set
and returns a new instance of `Grid` with the updated
multi-index set.

Additionally:

- The test suite has been updated to reflect the current changes.
- A new How-To Guide that demonstrates the method usage is added
  to the docs.

The commit should resolve Issue #158.