Consistent dimension expansion of polynomial instances.
The dimension expansion of a polynomial instance
to either a target dimension or to a target polynomial instance
is now supported by the `expand_dim()` method.
The procedure also makes sure that the underlying
user and internal domains are expanded accordingly.
Exceptions will be raised if they are incompatible.

Additionally:

- A new verification function `verify_poly_coeffs()`
  is added to the verification module to specifically
  verify polynomial coefficients.
- The test suite has been updated to reflect these changes.