Refactor polynomial-polynomial addition in the canonical basis.
Following the refactoring of the `Grid`
and `MultiIndexSet` classes, the polynomial-polynomial addition
in the canonical basis has been refactored.
The procedure is as follows: match the dimension
of the operands, unionize the grid, unionize the multi-index set,
sum up the coefficients according to the multi-index set,
and create a new instance.

The concrete implementation of the polynomial-polynomial subtraction
in the canonical basis has been removed in favor of using
the concrete implementation of addition operation and
the `__neg__()` dunder method (i.e., subtraction is an addition
between the first operand and the negation of the second operand).

The test suite has been updated to include polynomial-polynomial
addition in the canonical basis which is relegated to
the polynomial test module instead of canonical polynomial
specific test module.

This commit should resolve Issue #161.