Refactor the multiplication of polynomials in the Newton basis.
The multiplication of polynomials in the Newton basis now
employs the multiplication of respective grids and multi-index sets.
Furthermore, it can handle the multiplication of polynomials
whose indices are separate (the multi-index set of the polynomial
differs from the set of the grid of the polynomial).

The test suite has been extended to verify the new changes.

This commit should resolve Issue #165.