Refactor verification functions for polynomial evaluation input and output.
New verification functions, namely, `shape_eval_output()`
and `verify_query_points()` are introduced to streamline
the input/output verification process in the polynomial evaluation.

Additionally:

- The test suite has been updated to include the testing of
  the new changes.
- Issues due to batch evaluation in the Newton evaluation has
  been fixed.

This commit should resolve Issue #168, specifically,
the evaluation of a polynomial with multiple coefficient sets
on a single query point should return two-dimensional array
with a single row (the number of points) and multiple columns
(the number of coefficient sets).