Refactor polynomial-polynomial addition in the Newton basis.
Following the refactoring of the `Grid`
and `MultiIndexSet` classes, the polynomial-polynomial addition
in the Newton basis has been refactored.
The procedure is as follows: match the dimension
of the operands, unionize the grid, unionize the multi-index set,
compute the Lagrange coefficients of the operands
at the unisolvent nodes, sum them up, transform to the Newton basis,
and finally create a new instance.

The test suite has been updated to include polynomial-polynomial
addition in the Newton basis which is delegated to
the polynomial test module instead of the Newton polynomial
specific test module.

This commit should resolve Issue #166.