Refactor polynomial-polynomial multiplication in the Lagrange basis.
The implementation of general polynomial-polynomial
multiplication in the Lagrange basis has been removed to avoid
unnecessary coupling with the Newton polynomial class and
the corresponding transformation.
LagrangePolynomial would be the most bare polynomial basis
in Minterpy.

The support for scalar and constant polynomial multiplication
remains because it does not require any transformation.

The test suite has been updated to verify the behavior of
the changes.

This commit should resolve Issue #169.