Add polynomial basis transformation methods to Interpolant class
Several methods to transform the underlying Newton polynomial
of an `Interpolant` instance to various polynomial bases
are introduced. This includes to Newton, Lagrange, canonical,
and Chebyshev.

The corresponding test suite has been updated to validate these methods.
Finally, the in-depth tutorials have been updated with reference
to the new methods.

This commit should resolve Issue #182.